Add function to update oAuth access token from Google
The Google API expects an access token as authorization when making a
service call. During the configuration of gCalSync, the initialization
process creates an access token. However, each token has a lifetime (by
default of 3600 seconds). If it expires, the API calls won't be able to
complete due to lack of authorization. To prevent this from happening, the
token has to be refreshed if it has.

This function should be called before any call using the Google API, to
guarantee that the token is always valid.