Add function to support retrieval of calendars from the Google Service
To insert an event via the API, we are required to pass a calendar ID
so that if the account we're associated to has multiple accounts, it
populates the intended calendar. This function will support retrieval
of the calendar list from the associated account, so that it can be
selected by the administrator on the gCalSync configuration page.