Add functionality to insert SMF Calendar entry into Google Calendar
There are two parts to this commit - the first is the code required to
perform the actual operation which is included in the gcalsync_insert
function. The second is the modification of the SMF code base to trigger
the new function on a calendar entry insert. The second piece is
facilitated by the package installer through parsing the changes defined
in gCalSync.xml.

The implementation works but the code added to the SMF base isn't optimal.
There's room for improvement, especially in the portions that I know
are going to be redundant in the remove and update functions.