Add functionality to update SMF Calendar entry with assoc to Google Calendar
Same goes for this code, as the gcalsync_insert function. It's made up
of two parts; one in gCalSync.php and the other as a change to the SMF code

The implementation works but the code added to the SMF base isn't optimal.
There's room for improvement, as I've re-used some of the code in the SMF
insert function to initialize the gCalSync service in the update call.