Add functionality to remove Google Calendar entry when triggered by SMF
The logic to perform this operation is all of 10 lines. The integration
into SMF is what caused the heartburn. Through testing, I discovered
that if an event is created with a post, the logic to update and delete
that event is in a COMPLETELY different place than the rest of the event
management logic within the SMF code base.